chore: move kbMatchStatements value getters to common place DEVSU-2574
